Levelling up. Modifiers. Lazy init.
Added particle effect for levelling up.

Switched to current value/max value format for health and xp UI.

Added IModifierProvider interface to allow for damage buffs (and in the future other stat buffs). Currently, weapons buff damage depending on the level. The modifier is either additive or calculated as a percentage.

Introduced lazy initialization for values in Health, AIController, BaseStats and Fighter.