Camera Controller
Removed CinemachineCamera which followed the player character.
Implemented an input-based camera mechanic, where the camera position, rotation and zoom is controlled based on keyboard and mouse input + the option to toggle between following the player character and moving the camera independently.

TO DO: Clamp values for input based camera movement.