Inventory: InventoryItem and player's "backpack".
Inventory.cs on the player responsible for handling all changes to the inventory.
InventoryUI subscribes to the changes in the inventory and redraws the UI.
Currently the items that show up in the inventory are hard-coded in for debugging purposes. They function as regular pickups would and can be moved around the inventory.
Inventory items are scriptable objects with unique IDs. They are loaded from Resources when the game first loads.