Inventories: Loot
Changing the way pickups and enemy drops are displayed. Instead of physical pickups, items that can be added to inventory will mostly be shown as their UI representations which are displayed when the player clicks on a chest/container or a dead enemy.
Keeping the old way as well because it is needed for when the player drops an item from inventory and wants to pick it up again.